Admin features

+++ Backend/src/main/java/com/example/demo/repository/FoundItemRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.FoundItem;
 
-public interface FoundItemRepository extends JpaRepository<FoundItem, Integer> {}
+public interface FoundItemRepository extends JpaRepository<FoundItem, Long> {}
