+++ Backend/src/main/java/com/example/demo/repository/LostItemRepository.java
@@ -5,4 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.LostItem;
 
-public interface LostItemRepository extends JpaRepository<LostItem, Integer> {}
+public interface LostItemRepository extends JpaRepository<LostItem, Long> { }
+
